# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RenameCluster": {
        "additionalProperties": false,
        "properties": {
          "cluster_name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Cluster Name",
            "type": "string"
          }
        },
        "required": [
          "cluster_name"
        ],
        "title": "RenameCluster",
        "type": "object"
      },
...
      "RetirementPut": {
        "additionalProperties": false,
        "properties": {
          "cluster_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Id"
          },
          "retire_after_days": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retire After Days"
          },
          "warn_days": {
            "exclusiveMinimum": 0.0,
            "title": "Warn Days",
            "type": "number"
          }
        },
        "required": [
          "retire_after_days",
          "warn_days"
        ],
        "title": "RetirementPut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "JAVV backend",
    "version": "snapshot"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/clusters": {
      "get": {
        "description": "Retired clusters (issue 765) are left out unless asked for: the switcher and All clusters\nboth read this list, so leaving one out here is what takes it off them. Each row carries the\nnewest accepted scan (`last_scan_at`) and its retirement schedule for the warning banner:\n`silent_since` is where the countdown starts (the newest scan, else the first token's mint,\nor the return from retirement if later), then `warns_at` and `retires_at` (null = never).",
        "operationId": "list_clusters_api_v1_clusters_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_retired",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Retired",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Clusters Api V1 Clusters Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Clusters",
        "tags": [
          "clusters"
        ]
      }
    },
    "/api/v1/clusters/{cluster_id}/name": {
      "put": {
        "operationId": "rename_cluster_api_v1_clusters__cluster_id__name_put",
        "parameters": [
          {
            "in": "path",
            "name": "cluster_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9-]{7,63}$",
              "title": "Cluster Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameCluster"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Rename Cluster Api V1 Clusters  Cluster Id  Name Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Cluster",
        "tags": [
          "clusters"
        ]
      }
    },
    "/api/v1/clusters/{cluster_id}/retire": {
      "post": {
        "description": "Manual retire: off the list, tokens revoked, data kept (issue 765).",
        "operationId": "retire_api_v1_clusters__cluster_id__retire_post",
        "parameters": [
          {
            "in": "path",
            "name": "cluster_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9-]{7,63}$",
              "title": "Cluster Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Retire Api V1 Clusters  Cluster Id  Retire Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retire",
        "tags": [
          "clusters"
        ]
      }
    },
    "/api/v1/clusters/{cluster_id}/unretire": {
      "post": {
        "operationId": "unretire_api_v1_clusters__cluster_id__unretire_post",
        "parameters": [
          {
            "in": "path",
            "name": "cluster_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9-]{7,63}$",
              "title": "Cluster Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Unretire Api V1 Clusters  Cluster Id  Unretire Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unretire",
        "tags": [
          "clusters"
        ]
      }
    },
...
    "/api/v1/settings/retirement": {
      "get": {
        "description": "The EFFECTIVE window for a cluster, and whether it is that cluster's own override.",
        "operationId": "get_retirement_api_v1_settings_retirement_get",
        "parameters": [
          {
            "in": "query",
            "name": "cluster_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cluster Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Retirement Api V1 Settings Retirement Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Retirement",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "put_retirement_api_v1_settings_retirement_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetirementPut"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Put Retirement Api V1 Settings Retirement Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Retirement",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}